+++ CHANGELOG.md
@@ -3,18 +3,21 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Changelog](http://keepachangelog.com/).
 
-
 ## Unreleased
+
 ---
 
 ### New
 
+- [#838](https://gitlab.com/meltano/meltano/issues/838) Add indicator for speed run plugins
+
 ### Changes
+
+- [#838](https://gitlab.com/meltano/meltano/issues/838) Speed run plugins prioritized to top of the list
 
 ### Fixes
 
 ### Breaks
-
 
 ## 0.38.0 - (2019-08-21)
 
@@ -35,7 +38,6 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - [#746](https://gitlab.com/meltano/meltano/issues/746) Prevent duplicate schedule, report, and dashboard creation if there is an existing item
 - [#900](https://gitlab.com/meltano/meltano/issues/900) Fix fallback value for Iso8601 dates/times
 - [#903](https://gitlab.com/meltano/meltano/issues/903) Fix columns display issue for the base table in Analyze
-
 
 ## 0.37.2 - (2019-08-19)